+++ apps/admin/src/TableSearchMeta/Role/RoleDetailsMeta.tsx
@@ -7,7 +7,7 @@ import { UPDATE_SUCCESSFULLY } from "~/Constants"
 import { MetaDrivenFormModalOpenButton } from "~/packages/components/Modal/MetaDrivenFormModal/MetaDrivenFormModalOpenButton"
 // import { RoleFormMeta } from "~/Component/Feature/Roles/FormMeta/RoleFormMeta"
 import { REFRESH_PAGE } from "~/packages/utils/EventBus"
-import { getRoleUpdateFormMeta } from '~/Component/Feature/Roles/FormMeta/RoleUpdateFormMeta'
+import { getRoleFormMeta } from '~/Component/Feature/Roles/FormMeta/RoleFormMeta'
 
 export const getRoleDetailsMeta = (role: { [key: string]: any }): IDetailsMeta => {
   const updateEntity = QueryConstructor(((data) => RoleQueries.update({ ...data, params: { id: role.id } }).then(resp => {
@@ -22,7 +22,7 @@ export const getRoleDetailsMeta = (role: { [key: string]: any }): IDetailsMeta =
     cardActions: [
       <MetaDrivenFormModalOpenButton
         formTitle={`Update ${role.name}`}
-        formMeta={getRoleUpdateFormMeta(role.id)}
+        formMeta={getRoleFormMeta(role.id)}
         initialFormValue={{ name: role.name, permissions: role.permissions.map((permission: any) => permission.id), menu_permissions: role.menu_permissions }}
         formSubmitApi={updateEntity}
         buttonLabel={`Update Role`}
